sensor/requirements.txt
gpiozero
requests
python-dotenv
